bump version to v3
