+++ _examples/collectibles/main.go
@@ -10,8 +10,8 @@ import (
 	"os"
 	"time"
 
-	"github.com/fox-one/mixin-sdk-go/v2"
-	"github.com/fox-one/mixin-sdk-go/v2/mixinnet"
+	"github.com/fox-one/mixin-sdk-go/v3"
+	"github.com/fox-one/mixin-sdk-go/v3/mixinnet"
 	"github.com/gofrs/uuid"
 )
