+++ _examples/nodemonitor/monitor.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"time"
 
-	"github.com/fox-one/mixin-sdk-go/v2/mixinnet"
+	"github.com/fox-one/mixin-sdk-go/v3/mixinnet"
 	"github.com/sirupsen/logrus"
 )
