+++ address_mix.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	"github.com/btcsuite/btcd/address/v2/base58"
-	"github.com/fox-one/mixin-sdk-go/v2/mixinnet"
+	"github.com/fox-one/mixin-sdk-go/v3/mixinnet"
 	"github.com/gofrs/uuid/v5"
 )
