+++ tip.go
@@ -3,7 +3,7 @@ package mixin
 import (
 	"crypto/sha256"
 
-	"github.com/fox-one/mixin-sdk-go/v2/mixinnet"
+	"github.com/fox-one/mixin-sdk-go/v3/mixinnet"
 )
 
 const (
